First auto-updater implementation
At server launch, a thread is launched which performs an update each 20 seconds (for now, will be moved to a server config or ENV var.

Moved datalad_update to the the utils module to group all functionalities.

The exact way to handle the auto_updater thread closing is still opened to discussion.