Fail safe publish
In case a publish fails, a new thread is launched which repeatedly attempts to publish the new changes to the remote. An option for this thread was added to allow users to drop content upon publish completion. However, dropping content will very probably become a more complex issue if files are sent locally to Boutiques instead of the cluster downloading them directly.