Safety save on server startup
When starting the server, for users that do not currently have an execution running, we will datalad save their data folder. This will insure that if some files had been written to their folder, but not successfully saved on datalad before the server shut down, it will now save these files for them to be published by the auto publisher. We do not do this for user folders where executions are still "Running" as the executions might need these files unlocked to edit them (saving them will make them read only) and after the execution finishes, it will save the user folder on its own.