Fix Execution Kill + Path safe validation change
On execution kill, we now properly save the created files for the killed execution.
Changed is_data_accessible for when using a dataset. We follow the symlink until we reach the actual file. If the file is part of the .git/annex/objects, we use the absolute of the last symlink that linked to it to validate if user has access.