Util functions to follow datalad symlinks
In case we are using a dataset, we want to get the final path that links to git-annex. As such for paths that are symlinks, we will follow symlinks until the file referenced. If this file is in the git-annex objects folder, we will return the last symlink linking to it. This mimics the os.path.exists functionning.