Remove all mentions of datalad_drop
The drop should not be performed by a user. As an unlocked unchanged file will actually be drop successfully, this may cause issues during executions. We may unlock a file to allow an execution to edit it for example, but by dropping it, this would fail and cause issues. When using git-annex drop --key, we can drop the file from the .git/annex/objects without the unlock copy being dropped. We are still not imuned from someone editing a file in the meantime, but this is still a much better solution.