Improved datalad_get and archive creation
Regarding Datalad get, from the given path, we now explore symlinks in that folder to go and get their actual content. This is done iteratively instead of recursively to make sure no stack overflow will take place. When adding new paths, we also make sure that no path parent to that one were already added to the list as this other get will be included in the parent get. This also insures circular symlinks will be managed and won't cause the system to iterate forever.

Regarding archive creation, we now also traverse files and folders. For files, if they are symlinks and their content is accessible by the user, we will return the file with its actual content instead of a symlink. Otherwise, the symlink itself will be returned. For folders, if a symlink links to a folder, it will be returned as a symlink. As we add files manually one by one, we must also add empty folders to the archive as those will otherwise not be added to the archive.