Fixing symlink within folders accessed through symlinks
As when a folder is accessed through a symlink, the path simply includes the symlink (even when using normpath), relative symlinks can be uncorrectly resolved as they are not relative to the correct path anymore. To fix this, before calling normpath, a realpath is called on the parent folder. Improvement to accelerate find_paths_to_get for files.