Changes to /path PUT to manage files better
Instead of getting the entire possible folder where a file would be written, we now pin points the files to be overwritten when using datalad. For an archive, we only get and unlock existing files that will be overwritten by the archive. This greatly limits the amount of files that are uselessly gotten.