Using setup config when launching server with CLI
When using the CLI without Docker, the config created with `carmin-server setup` was not used. It is now used if it exists.
Additionnally, when initializing the server, the validation for the existence of the pipeline and data folders were performed after the supported descriptors folder creation which would lead to poorly documented crashed in case the pipelines folder does not exists.